+++ ring/requirements.txt
@@ -186,6 +186,8 @@ mypy-boto3-sqs==1.34.121
     # via boto3-stubs
 nodeenv==1.9.1
     # via pre-commit
+numpy==2.2.5
+    # via pgvector
 orjson==3.9.10
     # via fastapi
 packaging==25.0
@@ -202,6 +204,8 @@ pgcli==4.0.1
     # via ring (pyproject.toml)
 pgspecial==2.1.1
     # via pgcli
+pgvector==0.4.1
+    # via ring (pyproject.toml)
 platformdirs==4.3.6
     # via virtualenv
 pluggy==1.5.0
