+++ ring/tests/dev_requirements.txt
@@ -189,6 +189,8 @@ mypy-boto3-sqs==1.38.0
     # via boto3-stubs
 nodeenv==1.9.1
     # via pre-commit
+numpy==2.2.5
+    # via pgvector
 orjson==3.10.18
     # via fastapi
 packaging==25.0
@@ -203,6 +205,8 @@ pgcli==4.3.0
     # via ring (pyproject.toml)
 pgspecial==2.2.1
     # via pgcli
+pgvector==0.4.1
+    # via ring (pyproject.toml)
 platformdirs==4.3.7
     # via virtualenv
 pluggy==1.5.0
